Added container / image async APIs
