+++ src/core.rs
@@ -1,11 +1,15 @@
 pub use self::container::Container;
+pub use self::container_async::ContainerAsync;
 pub use self::docker::{Docker, Logs, Ports, RunArgs};
-pub use self::docker_async::{ContainerAsync, DockerAsync};
+pub use self::docker_async::DockerAsync;
 pub use self::image::{Image, Port};
+pub use self::image_async::ImageAsync;
 pub use self::wait_for_message::{WaitError, WaitForMessage};
 
 mod container;
+mod container_async;
 mod docker;
 mod docker_async;
 mod image;
+mod image_async;
 mod wait_for_message;
